requirements.txt
openai
httpx
python-dotenv
langgraph
langchain_core
langchain_openai
langchain_community
psycopg2
psycopg-pool
langchain_google_genai
langgraph-checkpoint-postgres
psycopg[binary]
sqlalchemy
uvicorn
fastapi[standard]
langgraph-cli[inmem]
faiss-cpu
qdrant-client
langgraph-qdrant
pydantic
asyncpg

